# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Agronom Bot Internal API
  version: "1.10.0"
  description: |
    Telegram-бот «Карманный агроном» — API v1.10.0.
    Fixes: +signature, +UNAUTHORIZED, +monthly limits, +X-API-Ver usage, +multipart required, clarified dosage units, +X-User-ID header, +ask_expert endpoint, +protocol fields category/status/waiting_days, +assistant chat/confirm_plan.
  contact:
    name: Ivan Gromov
    url: https://github.com/agronom-bot
    email: i.gromov@agrostore.ru
servers:
  - url: https://api.agronom.local
    description: Production
  - url: https://staging.agronom.local
    description: Staging
security:
  - ApiKeyAuth: []
    RequestSignAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    RequestSignAuth:
      type: apiKey
      in: header
      name: X-Req-Sign
    HmacAuth:
      type: apiKey
      in: header
      name: X-Sign
  parameters:
    ApiVersionHeader:
      name: X-API-Ver
      in: header
      required: true
      schema:
        type: string
        enum: [v1]
    UserIdHeader:
      name: X-User-ID
      in: header
      required: true
      schema:
        type: integer
    RequestTimestampHeader:
      name: X-Req-Ts
      in: header
      required: true
      schema:
        type: integer
        description: Unix timestamp (seconds)
    RequestNonceHeader:
      name: X-Req-Nonce
      in: header
      required: true
      schema:
        type: string
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key for payment creation retries.
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [NO_LEAF, LIMIT_EXCEEDED, GPT_TIMEOUT, BAD_REQUEST, UNAUTHORIZED, UPGRADE_REQUIRED, TOO_MANY_REQUESTS, SERVICE_UNAVAILABLE, FORBIDDEN]
        message:
          type: string
...
    AssistantChatRequest:
      type: object
      required: [message]
      properties:
        session_id:
          type: string
          example: "chat-6f7c"
        object_id:
          type: integer
          nullable: true
        message:
          type: string
          example: "Что делать с томатами после дождя?"
        metadata:
          type: object
          properties:
            recent_diagnosis_id:
              type: integer
              nullable: true
            plan_session_id:
              type: integer
              nullable: true
            locale:
              type: string
              example: ru
    AssistantProposal:
      type: object
      required: [proposal_id, kind]
      properties:
        proposal_id:
          type: string
        kind:
          type: string
          enum: [plan, event, clarify]
        plan_payload:
          type: object
          description: "Машинный план (тот же формат, что в PlanSession), draft/proposed."
          additionalProperties: true
        suggested_actions:
          type: array
          items:
            type: string
            enum: [pin, ask_clarification, show_plans, open_logbook]
    AssistantChatResponse:
      type: object
      required: [assistant_message]
      properties:
        assistant_message:
          type: string
          example: "Предлагаю обработку для 🍅 Грядка №2 и напомню за час."
        followups:
          type: array
          items:
            type: string
        proposals:
          type: array
          items:
            $ref: '#/components/schemas/AssistantProposal'
    AssistantConfirmRequest:
      type: object
      required: [proposal_id, object_id]
      properties:
        proposal_id:
          type: string
        object_id:
          type: integer
        preferred_time:
          type: string
          format: date-time
          nullable: true
        plan_session_id:
          type: integer
          nullable: true
    AssistantConfirmResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, scheduled]
        plan_id:
          type: integer
          nullable: true
        event_ids:
          type: array
          items:
            type: integer
        reminder_ids:
          type: array
          items:
            type: integer
...
paths:
  /v1/assistant/chat:
    post:
      summary: Conversational assistant entrypoint
      operationId: assistantChat
      description: "Получить ответ живого ассистента; чат — интерфейс, данные остаются в cases/plans/events."
      tags: [assistant]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantChatRequest'
      responses:
        '200':
          description: Assistant reply with proposals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantChatResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/assistant/confirm_plan:
    post:
      summary: Persist assistant proposal into plan/events/reminders
      operationId: assistantConfirmPlan
      description: "По клику «📌 Зафиксировать» создаёт/обновляет plan + события/напоминания через те же сервисы, что мастер."
      tags: [assistant]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantConfirmRequest'
      responses:
        '200':
          description: Plan/events persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantConfirmResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...